Stash

+++ designer/server/src/routes/forms/editor-v2/question-or-page-delete.js
@@ -1,8 +1,10 @@
+import { ComponentType } from '@defra/forms-model'
 import { StatusCodes } from 'http-status-codes'
 
 import * as scopes from '~/src/common/constants/scopes.js'
 import { deletePage } from '~/src/lib/editor.js'
 import * as forms from '~/src/lib/forms.js'
+import { getComponentsOnPageFromDefinition } from '~/src/lib/utils.js'
 import * as viewModel from '~/src/models/forms/editor-v2/question-delete.js'
 import { editorv2Path } from '~/src/models/links.js'
 
@@ -47,7 +49,7 @@ export default [
   }),
 
   /**
-   * @satisfies {ServerRoute<{ Params: { slug: string, pageId: string | undefined }}>}
+   * @satisfies {ServerRoute<{ Params: { slug: string, pageId: string, questionId?: string }}>}
    */
   ({
     method: 'POST',
@@ -55,11 +57,30 @@ export default [
     async handler(request, h) {
       const { params, auth } = request
       const { token } = auth.credentials
-      const { slug, pageId } = params
+      const { slug, pageId, questionId } = params
 
       const metadata = await forms.get(slug, token)
+      let deletePageOrQuestion = 'page'
+      if (questionId) {
+        // If only one (non-guidance question) on the page, 'deleting the question' becomes 'deleting the page'
+        const definition = await forms.getDraftFormDefinition(
+          metadata.id,
+          token
+        )
+        const components = getComponentsOnPageFromDefinition(definition, pageId)
+        const nonGuidanceComponents = components.filter(
+          (comp, idx) => comp.type === ComponentType.Markdown && idx === 0
+        )
+        if (nonGuidanceComponents.length > 1) {
+          deletePageOrQuestion = 'question'
+        }
+      }
 
-      await deletePage(metadata.id, token, pageId)
+      if (deletePageOrQuestion === 'question') {
+        // await deleteQuestion(metadata.id, token, pageId, questionId)
+      } else {
+        await deletePage(metadata.id, token, pageId)
+      }
 
       // Redirect POST to GET
       return h.redirect(editorv2Path(slug, 'pages')).code(StatusCodes.SEE_OTHER)
